+++ crates/wdl-grammar/src/lexer/v1.rs
@@ -381,6 +381,9 @@ pub enum Token {
     /// The `else` keyword.
     #[token("else")]
     ElseKeyword,
+    /// The `env` keyword.
+    #[token("env")]
+    EnvKeyword,
     /// The `false` keyword.
     #[token("false")]
     FalseKeyword,
@@ -560,6 +563,7 @@ impl<'a> ParserToken<'a> for Token {
             Self::CallKeyword => SyntaxKind::CallKeyword,
             Self::CommandKeyword => SyntaxKind::CommandKeyword,
             Self::ElseKeyword => SyntaxKind::ElseKeyword,
+            Self::EnvKeyword => SyntaxKind::EnvKeyword,
             Self::FalseKeyword => SyntaxKind::FalseKeyword,
             Self::HintsKeyword => SyntaxKind::HintsKeyword,
             Self::IfKeyword => SyntaxKind::IfKeyword,
@@ -646,6 +650,7 @@ impl<'a> ParserToken<'a> for Token {
             Self::CallKeyword => "`call` keyword",
             Self::CommandKeyword => "`command` keyword",
             Self::ElseKeyword => "`else` keyword",
+            Self::EnvKeyword => "`env` keyword",
             Self::FalseKeyword => "`false` keyword",
             Self::HintsKeyword => "`hints` keyword",
             Self::IfKeyword => "`if` keyword",
@@ -1408,7 +1413,8 @@ task
 then
 true
 version
-workflow"#,
+workflow
+env"#,
         );
         let tokens: Vec<_> = lexer.map(map).collect();
         assert_eq!(tokens, &[
@@ -1486,6 +1492,8 @@ workflow"#,
             (Ok(VersionKeyword), 222..229),
             (Ok(Whitespace), 229..230),
             (Ok(WorkflowKeyword), 230..238),
+            (Ok(Whitespace), 238..239),
+            (Ok(EnvKeyword), 239..242),
         ],);
     }